Add files via upload
The crypto currency converter API converts the value of a crypto currency such as bitcoin into its value in US dollars. The currency converter API can also be generalized to include other currencies like the Euro or the British pound.  However, for purposes of this project, we chose to work with crypto currencies like bitcoin.